Check if user has existing data and load it

+++ src/dashboard.js
@@ -11,13 +11,16 @@ export function dashboard() {
   let themeToggle = document.createElement('button');
   themeToggle.innerHTML = 'Toggle Theme';
 
-  let { demoAccount, inbox } = createDemo();
-  let currentProject = inbox;
-  localStorage.clear();
+  if (localStorage.getItem('user') === null) {
+    let { demoAccount, inbox } = createDemo();
+    let currentProject = inbox;
+    localStorage.clear();
+  }
 
-  // let demoAccount = JSON.parse(localStorage.getItem('user'));
-  // console.log(demoAccount);
-  // let currentProject = demoAccount.projects[0];
+  if (localStorage.getItem('user') !== null) {
+  let demoAccount = JSON.parse(localStorage.getItem('user'));
+  let currentProject = demoAccount.projects[0];
+  }
 
   let addProject = addNewProject(sidebar, demoAccount, todos);
   projectsContainer.className = 'projects-container';
